back/gestion-livraison-django-main/requirements.txt
asgiref==3.11.0
Django==6.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.14.0
django-filter==24.3.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.2.2
setuptools>=65.0.0
